Refactor `image_to_tensor_data`: move `image_to_pixvec` logic into `pixeldepth_support` and rename to `image_to_pixeldepth_vec`.
